tidy up codes for source mode
